Fix type casting
sizeof returns a size_t, but *hopefully*, sizeof(long) shouldn't be causing overflow problems!

+++ libvideo/libv4lconvert/spca501.c
@@ -27,7 +27,7 @@ void v4lconvert_spca501_to_yuv420(const u8 *src, u8 *dst, u32 width, u32 height,
 		/* -128 - 127 --> 0 - 255 and copy first line Y */
 		unsigned long *ldst = (unsigned long *)(dst + i * width);
 
-		for (unsigned int j = 0; j < width; j += sizeof(long)) {
+		for (unsigned int j = 0; j < width; j += (unsigned int) sizeof(long)) {
 			*ldst = *lsrc++;
 			*ldst++ ^= 0x8080808080808080ULL;
 		}
@@ -37,14 +37,14 @@ void v4lconvert_spca501_to_yuv420(const u8 *src, u8 *dst, u32 width, u32 height,
 			ldst = (unsigned long *)(dst + (width * height * 5) / 4 + i * width / 4);
 		else
 			ldst = (unsigned long *)(dst + width * height + i * width / 4);
-		for (unsigned int j = 0; j < width/2; j += sizeof(long)) {
+		for (unsigned int j = 0; j < width/2; j += (unsigned int) sizeof(long)) {
 			*ldst = *lsrc++;
 			*ldst++ ^= 0x8080808080808080ULL;
 		}
 
 		/* -128 - 127 --> 0 - 255 and copy second line Y */
 		ldst = (unsigned long *)(dst + i * width + width);
-		for (unsigned int j = 0; j < width; j += sizeof(long)) {
+		for (unsigned int j = 0; j < width; j += (unsigned int) sizeof(long)) {
 			*ldst = *lsrc++;
 			*ldst++ ^= 0x8080808080808080ULL;
 		}
@@ -54,7 +54,7 @@ void v4lconvert_spca501_to_yuv420(const u8 *src, u8 *dst, u32 width, u32 height,
 			ldst = (unsigned long *)(dst + width * height + i * width / 4);
 		else
 			ldst = (unsigned long *)(dst + (width * height * 5) / 4 + i * width / 4);
-		for (unsigned int j = 0; j < width/2; j += sizeof(long)) {
+		for (unsigned int j = 0; j < width/2; j += (unsigned int) sizeof(long)) {
 			*ldst = *lsrc++;
 			*ldst++ ^= 0x8080808080808080ULL;
 		}
@@ -69,7 +69,7 @@ void v4lconvert_spca505_to_yuv420(const u8 *src, u8 *dst, u32 width, u32 height,
 		/* -128 - 127 --> 0 - 255 and copy 2 lines of Y */
 		unsigned long *ldst = (unsigned long *)(dst + i * width);
 
-		for (unsigned int j = 0; j < width*2; j += sizeof(long)) {
+		for (unsigned int j = 0; j < width*2; j += (unsigned int) sizeof(long)) {
 			*ldst = *lsrc++;
 			*ldst++ ^= 0x8080808080808080ULL;
 		}
@@ -79,7 +79,7 @@ void v4lconvert_spca505_to_yuv420(const u8 *src, u8 *dst, u32 width, u32 height,
 			ldst = (unsigned long *)(dst + (width * height * 5) / 4 + i * width / 4);
 		else
 			ldst = (unsigned long *)(dst + width * height + i * width / 4);
-		for (unsigned int j = 0; j < width/2; j += sizeof(long)) {
+		for (unsigned int j = 0; j < width/2; j += (unsigned int) sizeof(long)) {
 			*ldst = *lsrc++;
 			*ldst++ ^= 0x8080808080808080ULL;
 		}
@@ -89,7 +89,7 @@ void v4lconvert_spca505_to_yuv420(const u8 *src, u8 *dst, u32 width, u32 height,
 			ldst = (unsigned long *)(dst + width * height + i * width / 4);
 		else
 			ldst = (unsigned long *)(dst + (width * height * 5) / 4 + i * width / 4);
-		for (unsigned int j = 0; j < width/2; j += sizeof(long)) {
+		for (unsigned int j = 0; j < width/2; j += (unsigned int) sizeof(long)) {
 			*ldst = *lsrc++;
 			*ldst++ ^= 0x8080808080808080ULL;
 		}
@@ -103,7 +103,7 @@ void v4lconvert_spca508_to_yuv420(const u8 *src, u8 *dst, u32 width, u32 height,
 	for (unsigned int i = 0; i < height; i += 2) {
 		/* -128 - 127 --> 0 - 255 and copy first line Y */
 		unsigned long *ldst = (unsigned long *)(dst + i * width);
-		for (unsigned int j = 0; j < width; j += sizeof(long)) {
+		for (unsigned int j = 0; j < width; j += (unsigned int) sizeof(long)) {
 			*ldst = *lsrc++;
 			*ldst++ ^= 0x8080808080808080ULL;
 		}
@@ -113,7 +113,7 @@ void v4lconvert_spca508_to_yuv420(const u8 *src, u8 *dst, u32 width, u32 height,
 			ldst = (unsigned long *)(dst + (width * height * 5) / 4 + i * width / 4);
 		else
 			ldst = (unsigned long *)(dst + width * height + i * width / 4);
-		for (unsigned int j = 0; j < width/2; j += sizeof(long)) {
+		for (unsigned int j = 0; j < width/2; j += (unsigned int) sizeof(long)) {
 			*ldst = *lsrc++;
 			*ldst++ ^= 0x8080808080808080ULL;
 		}
@@ -123,14 +123,14 @@ void v4lconvert_spca508_to_yuv420(const u8 *src, u8 *dst, u32 width, u32 height,
 			ldst = (unsigned long *)(dst + width * height + i * width / 4);
 		else
 			ldst = (unsigned long *)(dst + (width * height * 5) / 4 + i * width / 4);
-		for (unsigned int j = 0; j < width/2; j += sizeof(long)) {
+		for (unsigned int j = 0; j < width/2; j += (unsigned int) sizeof(long)) {
 			*ldst = *lsrc++;
 			*ldst++ ^= 0x8080808080808080ULL;
 		}
 
 		/* -128 - 127 --> 0 - 255 and copy second line Y */
 		ldst = (unsigned long *)(dst + i * width + width);
-		for (unsigned int j = 0; j < width; j += sizeof(long)) {
+		for (unsigned int j = 0; j < width; j += (unsigned int) sizeof(long)) {
 			*ldst = *lsrc++;
 			*ldst++ ^= 0x8080808080808080ULL;
 		}
